Tracing refactoring.
