--- tracing.conf
@@ -1,4 +1,0 @@
-type: JAEGER
-config:
-    service_name: ttt
-    sampler_param: 1
